+++ symfem/elements/huang_zhang.py
@@ -1,7 +1,7 @@
-"""Fortin-Soulie elements on a triangle.
+"""Huang-Zhang elements on a triangle.
 
-This element's definition appears in https://doi.org/10.1002/nme.1620190405
-(Fortin, Soulie, 1973)
+This element's definition appears in https://doi.org/10.1007/s11464-011-0094-0 
+(Huang, Zhang, 2011)
 """
 
 import typing
@@ -11,12 +11,36 @@ import sympy
 from ..finite_element import CiarletElement
 from ..functionals import ListOfFunctionals, PointEvaluation
 from ..functions import FunctionInput
-from ..polynomials import polynomial_set_1d_ZH
 from ..references import Reference
+from ..symbols import AxisVariablesNotSingle, x
+from ..functions import ScalarFunction, VectorFunction
 
+def polynomial_set_1d_ZH(
+         dim: int, variant: int, variables: AxisVariablesNotSingle = x
+) -> typing.List[ScalarFunction]:
+    """Huang-Zhang ."""
+    order = 2;
+    if dim == 1:
+        raise RuntimeError('The Haung-Zhang polynomial space is not implemented for dimension {}'.format(dim))
+    if dim == 2:
+        if variant == 0:
+            return [
+                ScalarFunction(variables[0] ** i * variables[1] ** j)
+                for j in range(order + 1)
+                for i in range(order)
+            ]
+        elif variant == 1:
+            return [
+                ScalarFunction(variables[0] ** i * variables[1] ** j)
+                for j in range(order)
+                for i in range(order+1)
+            ]
 
-class FortinSoulie(CiarletElement):
-    """Fortin-Soulie finite element."""
+    if dim == 3:
+        raise RuntimeError('The Haung-Zhang polynomial space is not implemented for dimension {}'.format(dim))
+
+class HuangZhang(CiarletElement):
+    """Huang-Zhang finite element."""
 
     def __init__(self, reference: Reference, order: int, variant: str):
         """Create the element.
@@ -53,7 +77,7 @@ class FortinSoulie(CiarletElement):
             ]
 
             poly: typing.List[FunctionInput] = []
-            poly += polynomial_set_1d_ZH(reference.tdim00)
+            poly += polynomial_set_1d_ZH(reference.tdim,0)
         super().__init__(reference, order, poly, dofs, reference.tdim, 1)
 
     names = ["Huang-Zhang", "HZ"]
